Do not publish your api key and it is uneeded
You do not need your api-key to make requests to the endpoints you are requesting. I have worked with the bazaar endpoint for the hypixel api and you do not need to add your key as a param. Also I believe you are not suppose to publish your api-key! Please get a new one if you can by doing "/api new" on hypixel.